Store key in home directory config file
By default the key will now be stored in the home directory config.
If desired, the config path can be set explicitly from the command-line.

The key is being removed from the .lockgit directory to make it less
likely to get checked into source control.  So while you can put the
config file anywhere, it is strongly recommended not to store your keys
inside the git directory.

If a repo was created from v0.5.x, it will be migrated to the new
format automatically when lockgit is run.  The old keyfile will not be
removed to prevent a horrible accident.