Use a more secure hashing mechanism
Changes the way files are compared to the vault contents, so a vault key is required to compare a file to the vault.

This fixes a security issue due to the fact the hashes of the file are public.

The hashes are salted which prevents the use of precomputed rainbow tables, but if you checked in a simple password that will not stand up to a dictionary attack, an attacker could brute force possibilities against the hash.

Initially, I thought it would be neat to check if a file had changed without having the key loaded into the app, so a salted hash seemed reasonable; but this was also thinking that your secrets will not fall to dictionary attacks. That no longer seems like a good assumption, nor does there seem to be reason to not load the key if you have it.

This change does not cause any previously computed hashes to be recomputed, which would be stored in git history anyway. If your keys will lose to a dictionary attack, you would be encouraged to change them and update LockGit. The new secrets will have hashes that cannot be compared without the key.